+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 ## Version 1.2.0
 
   * Register - get\_structure now only includes dates
+  * Gift Aid:
+    * Get Payments
+    * Update Payments
+    * Get Payment/Member Data
+    * Update Payment/Member Data
 
 ## Version 1.0.6